V1.0 - Platform and items spawn bug fixed + HUD added.

+++ SpaceRunner/src/entities/Entity.java
@@ -13,7 +13,7 @@ public class Entity {
 	protected Rectangle2D.Float hit_box;
 	protected LevelGenerator level_gen;
 	
-	public Entity(int x, int y, LevelGenerator level_gen) {
+	public Entity(float x, float y, LevelGenerator level_gen) {
 		this.x = x;
 		this.y = y;
 		this.level_gen = level_gen;
